refactor: onError에서 view 관련 로직을 컴포넌트단으로 옮기기
