+++ src/hooks/queries/useAddCartItem.ts
@@ -1,12 +1,13 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { addCartItem } from '../../api';
 import { MUTATION_KEY, QUERY_KEY } from '../../constant/queryKey';
-import { useToast } from '../useToast';
-import { ERROR } from '@/constant/message';
 
-const useAddCartItem = () => {
+interface Props {
+  onError: () => void;
+}
+
+const useAddCartItem = ({ onError }: Props) => {
   const client = useQueryClient();
-  const { showToast } = useToast();
 
   const { mutate, isError, isSuccess } = useMutation({
     mutationKey: [MUTATION_KEY.ADD_CART_ITEM],
@@ -16,9 +17,7 @@ const useAddCartItem = () => {
         queryKey: [QUERY_KEY.CART_ITEMS],
       });
     },
-    onError: () => {
-      showToast({ message: ERROR.addProduct, duration: 3000 });
-    },
+    onError,
   });
 
   return {
