+++ src/hooks/queries/usePatchCartItemQuantity.ts
@@ -1,12 +1,13 @@
 import { patchCartItemQuantity } from '@/api';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { useToast } from '../useToast';
 import { QUERY_KEY } from '@/constant/queryKey';
-import { ERROR } from '@/constant/message';
 
-const usePatchCartItemQuantity = () => {
+interface Props {
+  onError: () => void;
+}
+
+const usePatchCartItemQuantity = ({ onError }: Props) => {
   const client = useQueryClient();
-  const { showToast } = useToast();
 
   const { mutate, isSuccess, isError } = useMutation({
     mutationFn: patchCartItemQuantity,
@@ -15,9 +16,7 @@ const usePatchCartItemQuantity = () => {
         queryKey: [QUERY_KEY.CART_ITEMS],
       });
     },
-    onError: () => {
-      showToast({ message: ERROR.patchCartItemQuantity, duration: 3000 });
-    },
+    onError,
   });
   return {
     patchCartItemQuantity: mutate,
