+++ src/hooks/useCartItemQuantity.ts
@@ -1,9 +1,20 @@
+import { ERROR } from '@/constant/message';
 import useDeleteCartItem from './queries/useDeleteCartItem';
 import usePatchCartItemQuantity from './queries/usePatchCartItemQuantity';
+import { useToast } from './useToast';
 
 const useCartItemQuantity = () => {
-  const { deleteCartItem } = useDeleteCartItem();
-  const { patchCartItemQuantity } = usePatchCartItemQuantity();
+  const { showToast } = useToast();
+  const { deleteCartItem } = useDeleteCartItem({
+    onError: () => {
+      showToast({ message: ERROR.deleteProduct, duration: 3000 });
+    },
+  });
+  const { patchCartItemQuantity } = usePatchCartItemQuantity({
+    onError: () => {
+      showToast({ message: ERROR.patchCartItemQuantity, duration: 3000 });
+    },
+  });
 
   const decreaseCartItemQuantity = (cartItemId: number, quantity: number) => {
     if (quantity === 1) {
